requirements.txt
Django>=2.1,<2.2
Pillow>=5.2,<5.3
requests>=2.19.0,<2.20.0
uWSGI==2.0.17.1
psycopg2-binary==2.7.5
django-debug-toolbar==1.9.1
django-admin-sortable2==0.6.19
django-reversion==2.0.13
gunicorn>=19.9,<19.10
ipdb==0.11
djangorestframework==3.8.2
lxml==4.2.4
Markdown==2.6.11